sha1_mb: extend max size to over 4GB for sha1
APIs are kept the same. Test case is also added under
tests/extended. Makefile is written in md5 extend
patch and will cover sha1 test as well.

+++ sha1_mb/sha1_ctx_avx.c
@@ -36,7 +36,7 @@
 #endif
 
 static inline void hash_init_digest(SHA1_WORD_T * digest);
-static inline uint32_t hash_pad(uint8_t padblock[SHA1_BLOCK_SIZE * 2], uint32_t total_len);
+static inline uint32_t hash_pad(uint8_t padblock[SHA1_BLOCK_SIZE * 2], uint64_t total_len);
 static SHA1_HASH_CTX *sha1_ctx_mgr_resubmit(SHA1_HASH_CTX_MGR * mgr, SHA1_HASH_CTX * ctx);
 
 void sha1_ctx_mgr_init_avx(SHA1_HASH_CTX_MGR * mgr)
@@ -215,9 +215,9 @@ static inline void hash_init_digest(SHA1_WORD_T * digest)
 	memcpy_fixedlen(digest, hash_initial_digest, sizeof(hash_initial_digest));
 }
 
-static inline uint32_t hash_pad(uint8_t padblock[SHA1_BLOCK_SIZE * 2], uint32_t total_len)
+static inline uint32_t hash_pad(uint8_t padblock[SHA1_BLOCK_SIZE * 2], uint64_t total_len)
 {
-	uint32_t i = total_len & (SHA1_BLOCK_SIZE - 1);
+	uint32_t i = (uint32_t) (total_len & (SHA1_BLOCK_SIZE - 1));
 
 	memclr_fixedlen(&padblock[i], SHA1_BLOCK_SIZE);
 	padblock[i] = 0x80;
